Moved test harness to internal and renamed requestparsing to headers
