--- testing/README.md
@@ -1,2 +1,0 @@
-The `testing` package provides a harness and other test utilities for verifying
-the behaviour of the Go standard library HTTP related packages.
