+++ tests/headers/referer_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package requestparsing
+package headers
 
 import (
 	"context"
@@ -21,7 +21,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 
-	"github.com/google/go-safeweb/testing/requesttesting"
+	"github.com/google/go-safeweb/internal/requesttesting"
 )
 
 func TestReferer(t *testing.T) {
